requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
numpy==1.26.4
pandas==2.1.4
opencv-python-headless==4.8.0.74
easyocr==1.7.1
pillow==10.2.0
jinja2==3.1.3
python-multipart
openpyxl
python-Levenshtein